all files

+++ backend/controllers/AuditController.php
@@ -2,6 +2,7 @@
 
 namespace backend\controllers;
 
+use common\models\LoginForm;
 use Yii;
 use backend\models\Audit;
 use backend\models\AuditSearch;
@@ -35,6 +36,7 @@ class AuditController extends Controller
      */
     public function actionIndex()
     {
+        if (!Yii::$app->user->isGuest) {
         $searchModel = new AuditSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -42,6 +44,12 @@ class AuditController extends Controller
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
+        }else{
+            $model = new LoginForm();
+            return $this->redirect(['site/login',
+                'model' => $model,
+            ]);
+        }
     }
 
     /**
@@ -51,9 +59,16 @@ class AuditController extends Controller
      */
     public function actionView($id)
     {
+        if (!Yii::$app->user->isGuest) {
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
+        }else{
+            $model = new LoginForm();
+            return $this->redirect(['site/login',
+                'model' => $model,
+            ]);
+        }
     }
 
     /**
@@ -63,12 +78,19 @@ class AuditController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->isGuest) {
         $model = new Audit();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
         } else {
             return $this->render('create', [
+                'model' => $model,
+            ]);
+        }
+        }else{
+            $model = new LoginForm();
+            return $this->redirect(['site/login',
                 'model' => $model,
             ]);
         }
@@ -82,12 +104,19 @@ class AuditController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->isGuest) {
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
         } else {
             return $this->render('update', [
+                'model' => $model,
+            ]);
+        }
+        }else{
+            $model = new LoginForm();
+            return $this->redirect(['site/login',
                 'model' => $model,
             ]);
         }
@@ -101,9 +130,16 @@ class AuditController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->isGuest) {
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+        }else{
+            $model = new LoginForm();
+            return $this->redirect(['site/login',
+                'model' => $model,
+            ]);
+        }
     }
 
     /**
